requirements.txt
annotated-types==0.6.0
anthropic==0.15.0
anyio==4.2.0
backoff==2.2.1
certifi==2024.2.2
charset-normalizer==3.3.2
distro==1.9.0
exceptiongroup==1.2.0
filelock==3.13.1
fsspec==2024.2.0
google-generativeai==0.5.4
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
huggingface-hub==0.20.3
idna==3.6
packaging==23.2
pydantic==2.6.1
pydantic_core==2.16.2
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.31.0
sniffio==1.3.0
tokenizers==0.15.1
tqdm==4.66.1
typing_extensions==4.9.0
urllib3==2.2.0
videodb==0.2.3
